fixing repo export to/import from external directory
