+++ app/src/main/res/values-fr/strings.xml
@@ -255,7 +255,6 @@
   <string name="add_otp">Ajouter OTP</string>
   <string name="otp_import_success">Configuration TOTP importée avec succès</string>
   <string name="otp_import_failure_generic">Impossible d\'importer la configuration TOTP</string>
-  <string name="exporting_passwords">Exportation des mots de passe…</string>
   <string name="invalid_filename_text">Le nom du fichier ne doit pas contenir \'/\', définir le répertoire parent</string>
   <string name="directory_hint">Dossier</string>
   <string name="new_folder_set_gpg_key">Définir la clé PGP pour le dossier</string>
