+++ app/src/main/res/values-gl/strings.xml
@@ -264,7 +264,6 @@ a app desde unha fonte de confianza, como a Play Store, Amazon Appstore, F-Droid
   <string name="otp_import_success">Importouse correctamente a configuración TOTP</string>
   <string name="otp_import_failure_generic">Fallou a importación da configuración TOTP</string>
   <string name="otp_import_failure_no_selection">Non se seleccionou ningún ficheiro de imaxe</string>
-  <string name="exporting_passwords">Exportando contrasinais…</string>
   <string name="invalid_gpg_id">Atopouse .gpg-id, pero contén un ID de chave non válido ou descoñecido.</string>
   <string name="short_gpg_id">Atopouse .gpg-id, pero contén un ID hex curto, ao que non damos soporte</string>
   <string name="invalid_filename_text">O nome do ficheiro non pode conter \'/\', establece un directorio superior</string>
