+++ app/src/main/res/values/strings.xml
@@ -299,7 +299,6 @@
   <string name="otp_import_success">Successfully imported TOTP configuration</string>
   <string name="otp_import_failure_generic">Failed to import TOTP configuration</string>
   <string name="otp_import_failure_no_selection">No image file was selected</string>
-  <string name="exporting_passwords">Exporting passwords…</string>
   <string name="invalid_gpg_id">Found .gpg-id, but it contains an invalid or unknown key ID.</string>
   <string name="empty_gpg_id">Found .gpg-id, but it is empty.</string>
   <string name="missing_gpg_id">No .gpg-id could be found.</string>
